+++ packages/pacer-devtools/src/index.ts
@@ -2,16 +2,9 @@
 
 import * as Devtools from './core'
 
-// Create a dummy class for production that does nothing
-class DummyPacerDevtoolsCore {
-  constructor() {}
-  mount() {}
-  unmount() {}
-}
-
-export const PacerDevtoolsCore: (typeof Devtools)['PacerDevtoolsCore'] =
+export const PacerDevtoolsCore =
   process.env.NODE_ENV !== 'development'
-    ? (DummyPacerDevtoolsCore as any)
+    ? Devtools.PacerDevtoolsCoreNoOp
     : Devtools.PacerDevtoolsCore
 
 export type { PacerDevtoolsInit } from './core'