+++ packages/react-pacer-devtools/src/ReactPacerDevtools.tsx
@@ -1,21 +1,10 @@
-import React, { useEffect, useRef, useState } from 'react'
+import { createReactPanel } from '@tanstack/devtools-utils/react'
 import { PacerDevtoolsCore } from '@tanstack/pacer-devtools'
+import type { DevtoolsPanelProps } from '@tanstack/devtools-utils/react'
 
-export interface PacerDevtoolsReactInit {
-  theme?: 'light' | 'dark'
-}
+export interface PacerDevtoolsReactInit extends DevtoolsPanelProps {}
 
-export const PacerDevtoolsPanel = (props?: PacerDevtoolsReactInit) => {
-  const devToolRef = useRef<HTMLDivElement>(null)
+const [PacerDevtoolsPanel, PacerDevtoolsPanelNoOp] =
+  createReactPanel(PacerDevtoolsCore)
 
-  const [devtools] = useState(() => new PacerDevtoolsCore({}))
-  useEffect(() => {
-    if (devToolRef.current) {
-      devtools.mount(devToolRef.current, props?.theme ?? 'dark')
-    }
-
-    return () => devtools.unmount()
-  }, [devtools, props?.theme])
-
-  return <div style={{ height: '100%' }} ref={devToolRef} />
-}
+export { PacerDevtoolsPanel, PacerDevtoolsPanelNoOp }