+++ packages/react-pacer-devtools/src/production.ts
@@ -1,9 +1,7 @@
 'use client'
 
-import * as Devtools from './ReactPacerDevtools'
-
-export const PacerDevtoolsPanel = Devtools.PacerDevtoolsPanel
+export { PacerDevtoolsPanel } from './ReactPacerDevtools'
 
 export type { PacerDevtoolsReactInit } from './ReactPacerDevtools'
 
-export { pacerDevtoolsPlugin } from './plugin-prod'
+export { pacerDevtoolsPlugin } from './plugin'