+++ packages/solid-pacer-devtools/src/SolidPacerDevtools.tsx
@@ -1,22 +1,9 @@
+import { createSolidPanel } from '@tanstack/devtools-utils/solid'
 import { PacerDevtoolsCore } from '@tanstack/pacer-devtools'
-import { createSignal, onCleanup, onMount } from 'solid-js'
+import type { DevtoolsPanelProps } from '@tanstack/devtools-utils/solid'
 
-export interface PacerDevtoolsSolidInit {
-  theme?: 'light' | 'dark'
-}
+const [PacerDevtoolsPanel, PacerDevtoolsPanelNoOp] =
+  createSolidPanel(PacerDevtoolsCore)
+export interface PacerDevtoolsSolidInit extends DevtoolsPanelProps {}
 
-export const PacerDevtoolsPanel = (props?: PacerDevtoolsSolidInit) => {
-  let devToolRef: HTMLDivElement | undefined
-  const [devtools] = createSignal(new PacerDevtoolsCore({}))
-  onMount(() => {
-    if (devToolRef) {
-      devtools().mount(devToolRef, props?.theme ?? 'dark')
-
-      onCleanup(() => {
-        devtools().unmount()
-      })
-    }
-  })
-
-  return <div style={{ height: '100%' }} ref={devToolRef} />
-}
+export { PacerDevtoolsPanel, PacerDevtoolsPanelNoOp }