+++ packages/solid-pacer-devtools/src/index.ts
@@ -1,14 +1,14 @@
-'use client'
-
 import * as Devtools from './SolidPacerDevtools'
+import * as plugin from './plugin'
 
-export const PacerDevtoolsPanel: (typeof Devtools)['PacerDevtoolsPanel'] =
+export const PacerDevtoolsPanel =
   process.env.NODE_ENV !== 'development'
-    ? function () {
-        return null
-      }
+    ? Devtools.PacerDevtoolsPanelNoOp
     : Devtools.PacerDevtoolsPanel
 
-export type { PacerDevtoolsSolidInit } from './SolidPacerDevtools'
+export const pacerDevtoolsPlugin =
+  process.env.NODE_ENV !== 'development'
+    ? plugin.pacerDevtoolsNoOpPlugin
+    : plugin.pacerDevtoolsPlugin
 
-export { pacerDevtoolsPlugin } from './plugin'
+export type { PacerDevtoolsSolidInit } from './SolidPacerDevtools'