+++ packages/solid-pacer-devtools/src/production.ts
@@ -1,9 +1,5 @@
-'use client'
-
-import * as Devtools from './SolidPacerDevtools'
-
-export const PacerDevtoolsPanel = Devtools.PacerDevtoolsPanel
+export { PacerDevtoolsPanel } from './SolidPacerDevtools'
 
 export type { PacerDevtoolsSolidInit } from './SolidPacerDevtools'
 
-export { pacerDevtoolsPlugin } from './plugin-prod'
+export { pacerDevtoolsPlugin } from './plugin'