+++ packages/pacer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/pacer
 
+## 0.15.4
+
+### Patch Changes
+
+- bump event client version ([#81](https://github.com/TanStack/pacer/pull/81))
+
+## 0.15.3
+
+### Patch Changes
+
+- update event client version to support react native ([#73](https://github.com/TanStack/pacer/pull/73))
+
 ## 0.15.2
 
 ### Patch Changes